Move some mixing code
